Backport to C++17
We backport most of the code to C++17, and reduce the number of standard
library headers we use. This will hopefully enable wider usage and
faster compile times (due to the reduced number of includes).

This backport is not entirely complete: we still need to remove the
dependence on __VA_OPT__, and we need to avoid having to use "using
enum" statements. The plan from here is to find a workaround to remove
the need to have a recursive macro. First, though, we'll write compile
time benchmarks in order to compare compile times before and after
changes.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>