Fix bug in impl::sort, impl::sort_dedup
The base case where there were two items wasn't being handled
properly. Prior to the fix, the code here:

    auto &a = values[0], b = values[1];
    if (less(b, a)) {
            impl::swap(a, b);
    }

Meant that `b` WASN'T a reference, and was instead declared by
value. So `a` was being swapped with a temporary variable `b`, and NOT
`values[1]`, as was the intent. This bug appeared in similar form in
sort_dedup.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>